+++ CHANGELOG.md
@@ -37,6 +37,11 @@ Main feature of that release is full TypeScript support insluding mongoose model
 [BREAKING] ESM only. No more commonJS. That help to fix a lot of bugs with tests and provides better development expirience
 [BREAKING] Mongoose v8. https://mongoosejs.com/docs/migrating_to_8.html
 
+### 5.0.0-beta.20
+
+[FIX] fix bug with Lock model index
+[NEW] BaseModel add Virtuals
+
 ### 5.0.0-beta.19
 
 [NEW] intruducing new model type. BaseModel. Features - simplifie works with typescript. And based on statics